Updated SAL Action to support better extensibility
 - Updated Action contract - disallowed construction
   of invalid instances (fail fast).

 - Removed JaxB annotations from actions in SAL package

 - Added sal.rest bundle which contains REST version of
   data in SAL
   - Added REST JaxB annotated version of Actions
   - Added transformers for SAL Action to SAL Rest Action

 - Reworked Actions property to use Action classes instead
   of bitmask

 - Introduced APIs and services based on Transformators
   to register new OFActions,
   their SAL Action counterparts and mapping between them.

 - Updated JUnit integration test for SwitchManager to use
   new Actions contract

+++ opendaylight/forwardingrulesmanager/api/src/main/java/org/opendaylight/controller/forwardingrulesmanager/FlowConfig.java
@@ -23,7 +23,6 @@ import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
 import org.opendaylight.controller.sal.action.Action;
-import org.opendaylight.controller.sal.action.ActionType;
 import org.opendaylight.controller.sal.action.Controller;
 import org.opendaylight.controller.sal.action.Drop;
 import org.opendaylight.controller.sal.action.Flood;
@@ -1367,7 +1366,7 @@ public class FlowConfig implements Serializable {
                         ActionType.SET_NW_TOS.toString() + "=(.*)").matcher(
                         actiongrp);
                 if (sstr.matches()) {
-                    actionList.add(new SetNwTos(Byte.parseByte(sstr.group(1))));
+                    actionList.add(new SetNwTos((int) Byte.parseByte(sstr.group(1))));
                     continue;
                 }
